Re-export .Types from the Numeric.Sundials
